Clear pit Clippy diagnostics

+++ src/codegen/cranelift/translate_call.rs
@@ -691,9 +691,7 @@ impl<M: Module> CraneliftCodegen<M> {
                         let size_val = builder.ins().iconst(types::I64, size);
                         let alloc_inst = builder.ins().call(local_alloc, &[size_val]);
                         let heap_ptr = builder.inst_results(alloc_inst)[0];
-                        let copy_size = builder
-                            .ins()
-                            .iconst(module.isa().pointer_type(), size as i64);
+                        let copy_size = builder.ins().iconst(module.isa().pointer_type(), size);
                         builder.call_memcpy(
                             module.isa().frontend_config(),
                             heap_ptr,
@@ -722,7 +720,7 @@ impl<M: Module> CraneliftCodegen<M> {
                             .copied()
                             .expect("missing __olive_alloc");
                         let local_alloc = module.declare_func_in_func(heap_id, builder.func);
-                        let size_val = builder.ins().iconst(types::I64, size as i64);
+                        let size_val = builder.ins().iconst(types::I64, size);
                         let alloc_inst = builder.ins().call(local_alloc, &[size_val]);
                         let heap_ptr = builder.inst_results(alloc_inst)[0];
                         if results.len() == 1 {
